+++ rollup.config.js
@@ -26,50 +26,50 @@ function serve () {
 
 export default [
 	// MAIN command
-	// {
-	// 	input: 'command/main.js',
-	// 	output: { sourcemap: false, format: 'iife', name: 'app', file: 'public/command/bundle.js' },
-	// 	plugins: [
-	// 		svelte( { dev: !production, css: css => { css.write( 'bundle.css' ); }, preprocess: preprocess() } ),
-	// 		resolve( { browser: true, dedupe: [ 'svelte' ] } ),
-	// 		commonjs(),
-	// 		!production && serve(),
-	// 		!production && livereload( 'public' ),
-	// 		production && terser(),
-	// 		json()
-	// 	],
-	// 	watch: { clearScreen: true }
-	// },
-	// // DISPHENOID
-	// {
-	// 	input: 'video/main.js',
-	// 	output: { sourcemap: false, format: 'iife', name: 'app', file: 'public/video/bundle.js' },
-	// 	plugins: [
-	// 		svelte( { dev: !production, css: css => { css.write( 'bundle.css' ); }, preprocess: preprocess() } ),
-	// 		resolve( { browser: true, dedupe: [ 'svelte' ] } ),
-	// 		commonjs(),
-	// 		!production && serve(),
-	// 		!production && livereload( 'public' ),
-	// 		production && terser(),
-	// 		json()
-	// 	],
-	// 	watch: { clearScreen: true }
-	// },
-	// // DASH
-	// {
-	// 	input: 'admin/main.js',
-	// 	output: { sourcemap: false, format: 'esm', name: 'app', dir: 'public/admin/' },
-	// 	plugins: [
-	// 		svelte( { dev: !production, css: css => { css.write( 'bundle.css' ); }, preprocess: preprocess() } ),
-	// 		resolve( { browser: true, dedupe: [ 'svelte' ] } ),
-	// 		commonjs(),
-	// 		!production && serve(),
-	// 		!production && livereload( 'public' ),
-	// 		production && terser(),
-	// 		json()
-	// 	],
-	// 	watch: { clearScreen: true }
-	// }
+	{
+		input: 'command/main.js',
+		output: { sourcemap: false, format: 'iife', name: 'app', file: 'public/command/bundle.js' },
+		plugins: [
+			svelte( { dev: !production, css: css => { css.write( 'bundle.css' ); }, preprocess: preprocess() } ),
+			resolve( { browser: true, dedupe: [ 'svelte' ] } ),
+			commonjs(),
+			!production && serve(),
+			!production && livereload( 'public' ),
+			production && terser(),
+			json()
+		],
+		watch: { clearScreen: true }
+	},
+	// DISPHENOID
+	{
+		input: 'video/main.js',
+		output: { sourcemap: false, format: 'iife', name: 'app', file: 'public/video/bundle.js' },
+		plugins: [
+			svelte( { dev: !production, css: css => { css.write( 'bundle.css' ); }, preprocess: preprocess() } ),
+			resolve( { browser: true, dedupe: [ 'svelte' ] } ),
+			commonjs(),
+			!production && serve(),
+			!production && livereload( 'public' ),
+			production && terser(),
+			json()
+		],
+		watch: { clearScreen: true }
+	},
+	// DASH
+	{
+		input: 'admin/main.js',
+		output: { sourcemap: false, format: 'esm', name: 'app', dir: 'public/admin/' },
+		plugins: [
+			svelte( { dev: !production, css: css => { css.write( 'bundle.css' ); }, preprocess: preprocess() } ),
+			resolve( { browser: true, dedupe: [ 'svelte' ] } ),
+			commonjs(),
+			!production && serve(),
+			!production && livereload( 'public' ),
+			production && terser(),
+			json()
+		],
+		watch: { clearScreen: true }
+	},
 	// JSON
 	{
 		input: 'json/main.js',